# BaseParser: Replace if else chains with switch statement as it is made for these kind of use cases