As the BaseParser contained a lot of if else chains that could be replaced with switch statements i made this change which also improved the performance reached in the JMH benchmark.

Additionally i made the following changes to be more explicit about parameter/variable mutation.
Mark unchanged variables and parameters final.

I know the numbers in the JHM Becnhmark can vary on different runs but these are the numbers
that i got when running the benchmark on my machine before and after the changes.

Before the changes:
| Benchmark | Mode | Cnt | Score  | Units |
|-------------|--------|-----|--------|------|
| MyBenchmark.testParserAfterburner | thrpt | 2 | 227244,048 | ops/s |
| MyBenchmark.testParserJackson | thrpt | 2 | 224497,876 | ops/s |
| MyBenchmark.testParserQson | thrpt | 2 | 250616,141 | ops/s |
| MyBenchmark.testWriterAfterburner | thrpt | 2 | 217963,077 | ops/s |
| MyBenchmark.testWriterJackson | thrpt | 2 | 189072,941 | ops/s |
| MyBenchmark.testWriterQson | thrpt | 2 | 275120,601 | ops/s |

After the changes:
| Benchmark | Mode | Cnt | Score  | Units |
|-------------|--------|-----|--------|------|
| MyBenchmark.testParserAfterburner | thrpt | 2 |252945,414 | ops/s |
| MyBenchmark.testParserJackson | thrpt | 2 |239401,257 | ops/s |
| MyBenchmark.testParserQson | thrpt | 2 | 265371,373 | ops/s |
| MyBenchmark.testWriterAfterburner | thrpt | 2 | 222294,858 | ops/s |
| MyBenchmark.testWriterJackson | thrpt | 2 | 199144,777 | ops/s |
| MyBenchmark.testWriterQson | thrpt | 2 |290155,833 | ops/s |