@kevin0x90 2021-04-24
Hi,

thanks for the feedback. I’m not urgent with these changes and just wanted to give back a bit to open source.