@patriot1burke 2021-04-24
On closer look, your before/after numbers for parsing are worse.  You have to have a certain number of branches before switch statements become more performant.  GeeksforGeeks is telling me that number is 5 or more.  In some cases something like isDigit is faster with using greater/less than comparisons than a switch.

Rejecting until it can be proven your fixes are faster.